Split admin page to renderer, sanitizer, settings
