+++ admin/partials/smaily-admin-woocommerce-abandoned-cart.php
@@ -1,4 +1,7 @@
 <?php
+/**
+ * @var Smaily_Admin\Renderer $this
+ */
 
 defined( 'ABSPATH' ) || exit;
